# cmd/api/docs/swagger.yaml
definitions:
...
  afip.ConsultaEstadoCOEM:
    properties:
      CODE:
        type: string
      CuitAlta:
        type: string
      Estado:
        type: string
      FechaEstado:
        type: string
      IdentificadorCOEM:
        type: string
      Motivo:
        type: string
    type: object
  afip.ConsultaNoAbordo:
    properties:
      Contenedor:
        type: string
      Cuit:
        type: string
      DescripcionNoAbordo:
        type: string
      Destinacion:
        type: string
      FechaNoAbordo:
        type: string
      IdentificadorCACE:
        type: string
      IdentificadorCOEM:
        type: string
      MotivoNoAbordo:
        type: string
      Tipo:
        type: string
      TipoNoAbordo:
        type: string
    type: object
  afip.ConsultaSolicitud:
    properties:
      Estado:
        type: string
      FechaEstado:
        type: string
      IdentificadorCACE:
        type: string
      IdentificadorCOEM:
        type: string
      NumeroSolicitud:
        type: string
      TipoSolicitud:
        type: string
    type: object
...
  dto.DummyResponse:
    properties:
      AppServer:
        type: string
      AuthServer:
        type: string
      DbServer:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    email: sehogas@gmail.com
    name: Sebastian Hogas
  description: Esta API Json Rest actua como proxy SOAP a los servicios web de ARCA.
  termsOfService: http://swagger.io/terms/
  title: API Rest Json proxy a los servicios web SOAP ARCA
  version: "1.0"
paths:
...
  /coemcons/dummy:
    get:
      description: Visualizar el estado del servicio web, del servicio de autenticación
        y de la base de datos de ARCA
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DummyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Muestra el estado del servicio
      tags:
      - Consultas de Comunicación de Embarque
  /coemcons/obtener-consulta-estados-coem:
    get:
      consumes:
      - application/json
      description: Obtener Consulta Estados COEM
      parameters:
      - description: Identificador de la caratula
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/afip.ConsultaEstadoCOEM'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener Consulta Estados COEM
      tags:
      - Consultas de Comunicación de Embarque
  /coemcons/obtener-consulta-no-abordo:
    get:
      consumes:
      - application/json
      description: Obtener Consulta No Abordo
      parameters:
      - description: Identificador de la caratula
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/afip.ConsultaNoAbordo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener Consulta No Abordo
      tags:
      - Consultas de Comunicación de Embarque
  /coemcons/obtener-consulta-solicitudes:
    get:
      consumes:
      - application/json
      description: Obtener Consulta de Solicitudes
      parameters:
      - description: Identificador de la caratula
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/afip.ConsultaSolicitud'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener Consulta de Solicitudes
      tags:
      - Consultas de Comunicación de Embarque
...
swagger: "2.0"